refactor(route-utils): Moved route-utils2.js file
